vertex-main/ml-service/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
requests
google-cloud-aiplatform
google-genai
python-dotenv